pyproject.toml
[project]
name = "ansible-playbooks"
version = "0.1.0"
description = "Ansible playbooks for installing and configuring various software with Molecule tests"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.2.0",
    "ansible-lint>=25.1.2",
    "molecule>=25.2.0",
    "molecule-plugins[podman]>=23.7.0",
    "pytest-testinfra>=10.1.1",
]
